Refactor task package to make mutations more visible
